+++ ilert/data_source_status_page.go
@@ -42,9 +42,9 @@ func dataSourceStatusPageRead(ctx context.Context, d *schema.ResourceData, meta 
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for status page with name '%s' to be read", searchName))
+				return resource.RetryableError(fmt.Errorf("waiting for status page with name '%s' to be read, error: %s", searchName, err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read a status page with name: %s", searchName))
+			return resource.NonRetryableError(fmt.Errorf("could not read a status page with name: %s, error: %s", searchName, err.Error()))
 		}
 
 		found := resp.StatusPage