+++ ilert/data_source_status_page_group.go
@@ -59,9 +59,9 @@ func dataSourceStatusPageGroupRead(ctx context.Context, d *schema.ResourceData, 
 		if err != nil {
 			if _, ok := err.(*ilert.RetryableAPIError); ok {
 				time.Sleep(2 * time.Second)
-				return resource.RetryableError(fmt.Errorf("waiting for status page group with name '%s' to be read", searchName))
+				return resource.RetryableError(fmt.Errorf("waiting for status page group with name '%s' to be read, error: %s", searchName, err.Error()))
 			}
-			return resource.NonRetryableError(fmt.Errorf("could not read a status page group with name: %s", searchName))
+			return resource.NonRetryableError(fmt.Errorf("could not read a status page group with name: %s, error: %s", searchName, err.Error()))
 		}
 
 		found := resp.StatusPageGroup